+++ src/interface.rs
@@ -1,7 +1,7 @@
 use core::fmt::Debug;
 
 pub use crate::i2c::I2CInterface;
-pub use crate::spi::SPIInterface;
+pub use crate::spi::{SPIBusInterface, SPIInterface};
 
 /// async interface
 #[cfg(feature = "async")]