+++ src/spi.rs
@@ -1,7 +1,9 @@
 #[cfg(feature = "blocking")]
-use embedded_hal::spi::SpiDevice;
+use embedded_hal::spi::{SpiBus, SpiDevice};
 #[cfg(feature = "async")]
-use embedded_hal_async::spi::SpiDevice;
+use embedded_hal_async::spi::{SpiBus, SpiDevice};
+
+use embedded_hal::digital::OutputPin;
 
 use crate::Interface;
 
@@ -11,7 +13,6 @@ use crate::Interface;
 ///
 /// Using this wrapper struct instead of just an `embedded_hal::i2c::SpiDevice` we can easily support both
 /// I2C and SPI devices in the same driver.
-#[allow(unused)]
 pub struct SPIInterface<SPI: SpiDevice> {
     /// SPI device
     spi: SPI,
@@ -29,16 +30,74 @@ impl<SPI: SpiDevice> SPIInterface<SPI> {
     }
 }
 
+/// SPI interface for the driver using an `embedded_hal::spi::SpiBus` instead of `embedded_hal::spi::SpiDevice`
+/// and an `embedded_hal::digital::OutputPin` (for CS)
+pub struct SPIBusInterface<SPI: SpiBus, CS: OutputPin> {
+    /// SPI bus
+    spi: SPI,
+    /// Chip select pin
+    cs: CS,
+}
+
+impl<SPI: SpiBus, CS: OutputPin> SPIBusInterface<SPI, CS> {
+    /// Create a new SPI interface from an SPI bus and a chip select pin
+    /// that implement the `SpiBus` and `OutputPin` traits respectively.
+    ///
+    /// # Arguments
+    /// * `spi` - SPI bus
+    /// * `cs` - Chip select pin
+    #[allow(unused)]
+    pub fn new(spi: SPI, cs: CS) -> Self {
+        Self { spi, cs }
+    }
+}
+
+#[cfg(feature = "async")]
+impl<SPI: SpiBus, CS: OutputPin> Interface for SPIBusInterface<SPI, CS> {
+    type Error = SPI::Error;
+
+    async fn write_read(&mut self, write: &[u8], read: &mut [u8]) -> Result<(), Self::Error> {
+        self.cs.set_low().ok();
+        let result = self.spi.transfer(read, write).await;
+        self.cs.set_high().ok();
+        result
+    }
+
+    async fn write(&mut self, data: &[u8]) -> Result<(), Self::Error> {
+        self.cs.set_low().ok();
+        let result = self.spi.write(data).await;
+        self.cs.set_high().ok();
+        result
+    }
+}
+
+#[cfg(feature = "blocking")]
+impl<SPI: SpiBus, CS: OutputPin> Interface for SPIBusInterface<SPI, CS> {
+    type Error = SPI::Error;
+
+    fn write_read(&mut self, write: &[u8], read: &mut [u8]) -> Result<(), Self::Error> {
+        self.cs.set_low().ok();
+        let result = self.spi.transfer(read, write);
+        self.cs.set_high().ok();
+        result
+    }
+
+    fn write(&mut self, data: &[u8]) -> Result<(), Self::Error> {
+        self.cs.set_low().ok();
+        let result = self.spi.write(data);
+        self.cs.set_high().ok();
+        result
+    }
+}
+
 #[cfg(feature = "async")]
 impl<SPI: SpiDevice> Interface for SPIInterface<SPI> {
     type Error = SPI::Error;
 
-    #[allow(unused_variables)]
     async fn write_read(&mut self, write: &[u8], read: &mut [u8]) -> Result<(), Self::Error> {
         self.spi.transfer(read, write).await
     }
 
-    #[allow(unused_variables)]
     async fn write(&mut self, data: &[u8]) -> Result<(), Self::Error> {
         self.spi.write(data).await
     }
@@ -48,12 +107,10 @@ impl<SPI: SpiDevice> Interface for SPIInterface<SPI> {
 impl<SPI: SpiDevice> Interface for SPIInterface<SPI> {
     type Error = SPI::Error;
 
-    #[allow(unused_variables)]
     fn write_read(&mut self, write: &[u8], read: &mut [u8]) -> Result<(), Self::Error> {
         self.spi.transfer(read, write)
     }
 
-    #[allow(unused_variables)]
     fn write(&mut self, data: &[u8]) -> Result<(), Self::Error> {
         self.spi.write(data)
     }